move abstract type to top file
